backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
sqlmodel==0.0.21
jinja2==3.1.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
pytest==8.2.2
httpx==0.27.2
PyMySQL==1.1.0
